DEPLOYMENTS_BRANCH is passed pipeline->job->task
This allows commits to go on a branch rather than free-float.
When it free-floats, it's not pushed, and we lose the
state file (`-state.json`), which makes subsequent deploys fail.

Passing the branch through pipeline down to task is sub-optimal. What
I'd really like is the job to default to `master` rather than plumbing
DEPLOYMENTS_BRANCH all the way down; however, I don't see a better way.

fixes:
```
[detached HEAD a71ad7a] Concourse CI automated BOSH deployer :airplane:
```