:beetle: Push BOSH `-state.json` on failed deploys
Previously the state files were only pushed for successful deploys.
This was a problem when a BOSH deployment failed after the new VM was
created: the record of the new VM was never recorded in the manifest,
and subsequent deploys would also fail.

There were two aspects to the fix:
- we used an `if` before the `bosh create-env`, which doesn't trigger an
abort when `set -e` is in effect. This allows us to _commit_ the state
file of the broken deploy.
- we used an `ensure` block to _push_ the commit was pushed
regardless of the task's success.

fixes: <https://ci.nono.io/teams/main/pipelines/BOSH/jobs/bosh-azure.nono.io/builds/42>