Split `github_deployments_key` from `deployments.yml`
This makes our `fly -t nono sp -p BOSH ...` less confusing.

Previously we passed in `deployments.yml` twice: once to be interpolated
as a yml file (`-l`, "flag that can be used for filling in template
values in configuration from a YAML file"), and once as a variable (`-v
deployments_yml=...`). It was confusing, passing in the same file twice
but using it in different ways.

And the only reason we used the `-l` option was for but a single key,
`github_deployments_key`, which lent itself to being split-out, so we
did.