🐞 `bosh-azure`'s internal address matches everyone else's
It's now 10.0.0.6; previously it was 10.0.0.5 for reasons which are
unclear to me.

This is important because the SANs of the certificates have 10.0.0.6 (we
use the same certificate for all our BOSH directors).

fixes: `/var/vcap/sys/log/credhub.log`
```
Error starting ApplicationContext. To display the auto-configuration report re-run your application with 'debug' enabled.
2017-12-04T01:31:07.461Z [main] .... ERROR --- SpringApplication: Application startup failed
org.springframework.web.client.ResourceAccessException: I/O error on GET request for "https://10.0.0.5:8443/.well-known/openid-configuration": Certificate for <10.0.0.5> doesn't match any of the subject
alternative names: [10.0.0.6]; nested exception is javax.net.ssl.SSLPeerUnverifiedException: Certificate for <10.0.0.5> doesn't match any of the subject alternative names: [10.0.0.6]
```