Comodo's PKI root certificate
We need our Commercial CA's root certificate to insert into our BOSH
director manifests if we're using CredHub and UAA and we want to use
CA-recognized certificates.

When the certificate is downloaded from Firefox, a CR LF is at the end
of each line (even when downloaded using macOS, which like all other
flavors of UNIX, use LF to designate end-of-line). We converted it to
LF-only by using vim's `:set ff=unix`, but we are not sure this is
necessary.

this is a requirement in order to fix:
```
Listing deployment 'concourse' instances infos:
  Performing request GET 'https://bosh-gce.nono.io:25555/deployments/concourse/instances?format=full':
    Performing GET request:
      Requesting token via client credentials grant: Performing request POST 'https://104.154.39.128:8443/oauth/token': Performing POST request: Retry: Post https://104.154.39.128:8443/oauth/token: x509: certificate signed by unknown authority

Exit code 1
```