BOSH Lite director is updated to Xenial
This introduces a number of extra certificates -- God, the
certificate-thing is hog-wild, and makes one yearn for simpler days.
I doubt I'm introducing any vulnerabilities by publishing them because
this director will always be behind a firewall. The same can't be said
for my other directors, but I'm not gonna worry about that now.