Templatize (sic) deployment of nginx + NTP + PowerDNS
We simplify our deployment of nginx + NTP + PowerDNS because we'll be deploying
two (AWS, Azure) of them, not one. Admittedly, for our present
deployment of 1 server (AWS), this complicates things, for whereas before we
had one manifest, now we have a script *and* a manifest.

We extract the differing four variables to be easily set within our script:
- IaaS (e.g. "aws")
- Instance Type (e.g. "t3.nano")
- External IP (e.g. "52.0.56.137")
- sslip.io hostname (e.g. "52-0-56-137")

It would be really nice to bring in our 3rd VM, the GCE VM, but that's a
slightly different beast in that it also runs *Concourse* (the server), so its
manifest is fairly different. We'll possibly address that at a later date.