<https://ci.nono.io> uses Certigo-issued cert
Let's Encrypt rate-limited us, and I couldn't wait a week.

Also, the Concourse spec file is cryptic, and it turns out that there
are _sub-properties_ for the property `tls.cert`. This wasn't part of
the BOSH spec, and I had to read the template file to figure out how to
configure it correctly. Very annoying. I must confess that if I, with
all my experience, have trouble with deploying Concourse with TLS, then
I'm astounded anyone can get this working. And `bpm` adds what I feel to
be an unnecessary layer of complexity.

fixes: `/var/vcap/sys/log/web/web.stderr.log`
```
2019/09/11 20:18:55 http: TLS handshake error from 73.189.219.4:58842: 429 urn:acme:error:rateLimited: Error creating new cert :: too many certificates already issued for exact set of domains: ci.nono.io: see https://letsencrypt.org/docs/rate-limits/
```