Disk size labels are meaningful: "1G", "2G" etc...
Previously we used "default", "large", and "larger", but such names were
devoid of meaning, so we use more meaningful names (though we did keep
"default" for backwards compatibility with existing manifests).

We have a range of disks, 1G through 6G, which will enable testing
multiple increasing disk resizes.

We chose incremements of 1G because that appears to be the quantum size
that GCP uses.