💵 Terraform GKE: 1 instance, not 3!
By changing the location to a zone instead of a region, I cut down the
number of instances from 3 to 1. The instances were in zones a, c, and
f. I don't want to waste that much money!

According to the
[docs](https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/container_cluster)

> location - (Optional) The location (region or zone) in which the
cluster master will be created, as well as the default node location. If
you specify a zone (such as us-central1-a), the cluster will be a zonal
cluster with a single cluster master. If you specify a region (such as
us-west1), the cluster will be a regional cluster with multiple masters
spread across zones in the region, and with default node locations in
those zones as well