k-v.io: GKE services (ClusterIP & LoadBalancer)
Although we expose the etcd peer port (2380) to the external world, we
do not expose the etcd client port (2379). It's not necessary, and I
think we're at the limit of 5 inbound rules on the load balancer before
Google charges me more.

We also include a livenessProbe. Fortunately, `nc`, needed for the
livenessProbe, is available on the etcd Docker container.