ns-azure has a persistent disk for `etcd`
Having a persistent data disk (i.e. Azure "managed disk") to store
`etcd` data will help when rebuilding the VMs. It _should_ ensure that
the entire `etcd` database is not re-downloaded after a rebuild of the
VM.

It _should_ also mean that we don't have to delete & re-add the node to
the cluster after a rebuild.