Pipeline regularly checks Concourse-Vault integration
Prior to this commit, our Concourse-Vault integration would break and
we wouldn't know. Usually this was because we rolled the k8s nodes, and
the Vault would be sealed until we manually unsealed it.

With this commit we check the integration daily, so that if Concourse
can't contact Vault, we'll soon discover it.