BOSH Director: remove NATS passwd
We remove the NATS user & passwd for the mbus; we worry it was causing
problems for the NATS 2.0 upgrade. The user & passwd no longer are set
in `bosh-deployment` for any IaaS with the exception of
`bosh-lite-docker`.

Although we didn't upgrade NATS for this one, we did bring in many of
new required properties.

- bosh-vsphere-creds.yml — no change, merely re-ordered