Re-IP BOSH Director 10.0.2.250 → 10.9.2.250
I re-IP'ed the BOSH Director. I had to regenerate the certs along the
way because sometimes the IP address is buried in the SANs.

I wanted to re-IP to keep my IPs within 10.9/16.

I also updated the Cloud Config so that generic (non-CF) deployments are
deployed to 10.9.2.0/24.

What can I say? It's a flex.