Dockerfile fedora-golang-bosh uses stock Golang
Previously we had custom-installed Golang because Fedora's was too old,
but now the reverse is true: our custom install's is too old. This
commit remedies that by installing the stock Golang, and it makes our
Dockerfile smaller.

We also lay the groundwork for creating arm64 versions of our Dockerfile
so that we can use it on our M1, M2 macs.

Drive-by: update our pipeline to point to the new oci-build-task
Concourse resource (previously was Alex Suraci's; now it's Concourse's).