Destroy deprecated ns-azure.sslip.io VM & infra
We don't need to pay Azure $20 / month to maintain an sslip.io
nameserver that I've taken out of rotation. Besides, I'm not
particularly fond of Azure, and since my new job isn't so
cloud-dependent, I don't need an Azure VM for testing.

The new DNS servers are on OVH and Hetzner.