vault.nono.io
Rather than using Hashicorp's Vault, which has a more encumbered
license, we experiment with the Linux Foundation's OpenBao.

vault.nono.io is put on the guest network, with a publicly-accessible
IPv6 address. Note: the DNS entry for vault.nono.io only resolves the
AAAA record (IPv6), no IPv4.

This is so that Caddy can acquire a Let's Encrypt certificate.

I've made the decision to open Vault to the internet-at-large even
though they suggest not doing it. I'd prefer to reach my vault from
anywhere.

This is a precursor to installing Concourse CI, which will be used to
create a product that measures the popularity of various DNS servers
(e.g. 1.1.1.1, 8.8.8.8, 9.9.9.9).

Signed-off-by: Brian Cunnie <brian.cunnie@gmail.com>